refactor old names t eaterslab produced by poohslab
