+++ android/app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "labs.pooh.mycanteen"
+        applicationId "labs.pooh.eaterslab"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
