+++ android/app/src/androidTest/java/labs/pooh/eaterslab/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package labs.pooh.mycanteen
+package labs.pooh.eaterslab
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("labs.pooh.mycanteen", appContext.packageName)
+        assertEquals("labs.pooh.eaterslab", appContext.packageName)
     }
 }
