+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="labs.pooh.mycanteen">
+    package="labs.pooh.eaterslab">
 
     <uses-feature android:name="android.hardware.camera2" android:required="false"/>
 
@@ -18,7 +18,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity 
-            android:name=".HelloSelectActivity"
+            android:name="labs.pooh.eaterslab.HelloSelectActivity"
             android:label="@string/select_action_title">
 
             <intent-filter>
@@ -29,11 +29,11 @@
         </activity>
 
         <activity
-            android:name=".MapSearchActivity"
+            android:name="labs.pooh.eaterslab.MapSearchActivity"
             android:label="@string/search_map_title" />
 
         <activity
-            android:name=".MainActivity"
+            android:name="labs.pooh.eaterslab.MainActivity"
             android:hardwareAccelerated="true"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"/>
