+++ android/settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='MyCanteen'
+rootProject.name='EatersLab'
 include ':app'
